+++ git_autoshare/core.py
@@ -40,50 +40,87 @@ def git_bin():
 def repos():
     hosts = load_hosts()
     for host, repos in hosts.items():
-        for repo, orgs in repos.items():
+        for repo, repo_data in repos.items():
+            if isinstance(repo_data, dict):
+                orgs = repo_data.get('orgs', [])
+                private = repo_data.get('private', False)
+            else:
+                orgs = repo_data
+                private = False
             repo_dir = os.path.join(cache_dir(), host, repo)
             orgs = [org.lower() for org in orgs]
-            yield host, orgs, repo.lower(), repo_dir
+            yield host, orgs, repo.lower(), repo_dir, private
 
 
 def shared_urls():
-    for host, orgs, repo, repo_dir in repos():
+    for host, orgs, repo, repo_dir, private in repos():
         for org in orgs:
             for suffix in ('', '.git'):
-                repo_url = 'https://%s/%s/%s%s' % (host, org, repo, suffix)
-                yield repo_url, host, org, repo, repo_dir
-                repo_url = 'https://git@%s/%s/%s%s' % (host, org, repo, suffix)
-                yield repo_url, host, org, repo, repo_dir
-                repo_url = 'http://%s/%s/%s%s' % (host, org, repo, suffix)
-                yield repo_url, host, org, repo, repo_dir
-                repo_url = 'ssh://git@%s/%s/%s%s' % (host, org, repo, suffix)
-                yield repo_url, host, org, repo, repo_dir
-                repo_url = 'git@%s:%s/%s%s' % (host, org, repo, suffix)
-                yield repo_url, host, org, repo, repo_dir
+                if not private:
+                    repo_url = 'https://%s/%s/%s%s' % \
+                        (host, org, repo, suffix)
+                    yield repo_url, host, org, repo, repo_dir, private
+                    repo_url = 'https://git@%s/%s/%s%s' % \
+                        (host, org, repo, suffix)
+                    yield repo_url, host, org, repo, repo_dir, private
+                    repo_url = 'http://%s/%s/%s%s' % \
+                        (host, org, repo, suffix)
+                    yield repo_url, host, org, repo, repo_dir, private
+                repo_url = 'ssh://git@%s/%s/%s%s' % \
+                    (host, org, repo, suffix)
+                yield repo_url, host, org, repo, repo_dir, private
+                repo_url = 'git@%s:%s/%s%s' % \
+                    (host, org, repo, suffix)
+                yield repo_url, host, org, repo, repo_dir, private
 
 
-def prefetch_one(host, orgs, repo, repo_dir, quiet):
+def git_remotes(repo_dir='.'):
+    remotes = subprocess.check_output([git_bin(), 'remote'],
+                                      cwd=repo_dir, universal_newlines=True)
+    for remote in remotes.split():
+        url = subprocess.check_output([git_bin(), 'remote', 'get-url', remote],
+                                      cwd=repo_dir, universal_newlines=True)
+        yield remote, url.strip()
+
+
+def prefetch_one(host, orgs, repo, repo_dir, private, quiet):
     if not os.path.exists(os.path.join(repo_dir, 'objects')):
         if not os.path.exists(repo_dir):
             os.makedirs(repo_dir)
         subprocess.check_call([git_bin(), 'init', '--bare'], cwd=repo_dir)
+    existing_remotes = dict(git_remotes(repo_dir))
     for org in orgs:
-        try:
-            subprocess.check_output([git_bin(), 'remote', 'remove', org],
-                                    stderr=subprocess.STDOUT, cwd=repo_dir)
-        except subprocess.CalledProcessError:
-            pass
-        repo_url = 'https://%s/%s/%s.git' % (host, org, repo)
+        if private:
+            repo_url = 'ssh://git@%s/%s/%s.git' % (host, org, repo)
+        else:
+            repo_url = 'https://%s/%s/%s.git' % (host, org, repo)
+        if org in existing_remotes:
+            existing_repo_url = existing_remotes[org]
+            if repo_url != existing_repo_url:
+                subprocess.check_call([
+                    git_bin(), 'remote', 'set-url', org, repo_url],
+                    cwd=repo_dir)
+            del existing_remotes[org]
+        else:
+            subprocess.check_call([
+                git_bin(), 'remote', 'add', org, repo_url],
+                cwd=repo_dir)
         if not quiet:
-            print("git-autoshare prefetching", repo_url, "in", repo_dir)
+            print("git-autoshare remote", org, repo_url, "in", repo_dir)
+    # remove remaining unneeded remotes
+    for existing_remote in existing_remotes:
+        if not quiet:
+            print("git-autoshare removing remote", existing_remote,
+                  "in", repo_dir)
         subprocess.check_call([
-            git_bin(), 'remote', 'add', org, repo_url], cwd=repo_dir)
-    fetch_cmd = [git_bin(), 'fetch', '-f', '--all', '--tags']
+            git_bin(), 'remote', 'remove', existing_remote],
+            cwd=repo_dir)
+    fetch_cmd = [git_bin(), 'fetch', '-f', '--all', '--tags', '--prune']
     if quiet:
         fetch_cmd.append('-q')
     subprocess.check_call(fetch_cmd, cwd=repo_dir)
 
 
 def prefetch_all(quiet):
-    for host, orgs, repo, repo_dir in repos():
-        prefetch_one(host, orgs, repo, repo_dir, quiet)
+    for host, orgs, repo, repo_dir, private in repos():
+        prefetch_one(host, orgs, repo, repo_dir, private, quiet)