+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changes
 .. ----------
 .. -
 
+1.0.0b1 (2018-01-07)
+--------------------
+- support for private repositories
+- better handling of remotes in prefetch
+- prefetch with --prun
+
 1.0.0a2 (2017-10-10)
 --------------------
 - first packaged version